+++ TestMaker/Lib/TaskTypes/SingleChoice.cs
@@ -8,7 +8,7 @@ namespace Lib.TaskTypes
         public List<string> Options { get; protected set; }
         public int RightAnswerIndex { get; protected set; }
 
-        public SingleChoice(string question, List<string> options, int rightAnswerIndex, int mark)
+        public SingleChoice(string question, List<string> options, int rightAnswerIndex, double mark)
             : base(question, options, mark)
         {
             if (options is null)
@@ -24,7 +24,7 @@ namespace Lib.TaskTypes
         }
 
         [JsonConstructor]
-        public SingleChoice(int time, string question, List<string> options, int rightAnswerIndex, int mark)
+        public SingleChoice(int time, string question, List<string> options, int rightAnswerIndex, double mark)
             : base(time, question, options, mark) 
         {
             if (options is null)
